`leaflet-adapter` `src` module consolidation
