+++ src/index.ts
@@ -1,8 +1,8 @@
-export * from './circle.js'
-export * from './coordinate-reference-system/epsg-3857.js'
-export * from './map/map.js'
-export * from './marker.js'
-export * from './polygon.js'
-export * from './popup.js'
-export * from './tile-layer/tile-layer-osm.js'
-export * from './tile-layer/tile-layer.js'
+export * from './leaflet-adapter/circle.js'
+export * from './leaflet-adapter/coordinate-reference-system/epsg-3857.js'
+export * from './leaflet-adapter/map/map.js'
+export * from './leaflet-adapter/marker.js'
+export * from './leaflet-adapter/polygon.js'
+export * from './leaflet-adapter/popup.js'
+export * from './leaflet-adapter/tile-layer/tile-layer-osm.js'
+export * from './leaflet-adapter/tile-layer/tile-layer.js'
