+++ src/leaflet-adapter/circle.ts
