+++ src/leaflet-adapter/marker.ts
