+++ src/leaflet-adapter/popup.ts
